Object orientated business application written in C#
This object-orientated application provides a business solution to the
fictional travel company 'Sublime Tours'. The application allows staff
at Sublime Tours to create a tour for passengers outlining the start and
end dates for the tour, maximum number of passengers, travel distance of
the tour, cost mark-up as well as set different kinds of cost items for
the tour.

Different cost items follow an object-orientated hierarchical structure
allowing cost items to be polymorphic i.e. respond to the same message
differently.